deployment file and dev copy
